Check Owner-Left-Time in header after reading from cache
After reading from cache, check if there is Owner-Left-Time in header,
if so, compares the Owner-Left-Time and the latest left time  with the
same owner(ip).

If the left time reading from header less then the latest left time, it means
that the content maybe have been purged, we should drop the content from cache.

Change-Id: I5acf74e3e1267bc8661ec26008b371a2b60b8a01
Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>